Move dropzone components
