+++ app/src/lib/commit/CommitDragItem.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
 	import { CommitDragActionsFactory } from '$lib/commits/dragActions';
-	import CardOverlay from '$lib/components/Dropzone/CardOverlay.svelte';
-	import Dropzone from '$lib/components/Dropzone/Dropzone.svelte';
+	import CardOverlay from '$lib/dropzone/CardOverlay.svelte';
+	import Dropzone from '$lib/dropzone/Dropzone.svelte';
 	import { getContext, maybeGetContextStore } from '$lib/utils/context';
 	import { RemoteCommit, Branch, Commit } from '$lib/vbranches/types';
 	import type { Snippet } from 'svelte';
