+++ app/src/lib/dropzone/CardOverlay.svelte
