+++ app/src/lib/dropzone/LineOverlay.svelte
